refactor(consensus): move check_shard_version_matches_flags to authority_service
