+++ crates/starfish/core/src/authority_service.rs
@@ -26,7 +26,7 @@ use crate::{
     block_header::{
         BlockHeaderAPI, BlockHeaderDigest, BlockRef, GENESIS_ROUND, ShardWithProof,
         ShardWithProofAPI, ShardWithProofV1, SignedBlockHeader, TransactionsCommitment,
-        VerifiedBlock, VerifiedOwnShard, VerifiedTransactions, check_shard_version_matches_flags,
+        VerifiedBlock, VerifiedOwnShard, VerifiedTransactions,
     },
     block_verifier::BlockVerifier,
     commit::{CommitAPI as _, CommitRange, TrustedCommit},
@@ -598,6 +598,34 @@ impl<C: CoreThreadDispatcher> AuthorityService<C> {
             }
         }
     }
+}
+
+/// Rejects a deserialized `ShardWithProof` whose variant does not match the
+/// local protocol-flag configuration. The flag is uniform across the network
+/// within an epoch, so any mismatch implies either a malicious peer or a
+/// misconfigured upgrade path. In the flag-OFF (raw V1 wire form) branch the
+/// check is a tautology — the deserializer always produces V1 — but it is
+/// called there for symmetry.
+pub(crate) fn check_shard_version_matches_flags(
+    shard: &ShardWithProof,
+    protocol_config: &iota_protocol_config::ProtocolConfig,
+) -> ConsensusResult<()> {
+    let fast_commit_sync = protocol_config.consensus_fast_commit_sync();
+    let variant_matches_flags = matches!(
+        (shard, fast_commit_sync),
+        (ShardWithProof::V1(_), false) | (ShardWithProof::V2(_), true)
+    );
+    if !variant_matches_flags {
+        let actual = match shard {
+            ShardWithProof::V1(_) => "V1",
+            ShardWithProof::V2(_) => "V2",
+        };
+        return Err(ConsensusError::WrongShardVersionForFlags {
+            actual,
+            fast_commit_sync,
+        });
+    }
+    Ok(())
 }
 
 #[async_trait]
@@ -4023,10 +4051,8 @@ mod tests {
 
     #[test]
     fn check_shard_version_matches_flags_when_fast_commit_sync_enabled() {
-        use crate::{
-            block_header::{ShardWithProof, check_shard_version_matches_flags},
-            error::ConsensusError,
-        };
+        use super::check_shard_version_matches_flags;
+        use crate::{block_header::ShardWithProof, error::ConsensusError};
         let mut config = iota_protocol_config::ProtocolConfig::get_for_max_version_UNSAFE();
         config.set_consensus_fast_commit_sync_for_testing(true);
